+++ src/util/harvesterImpls.js
@@ -1,11 +1,10 @@
-const extractHarvesterImpls = (resources) => {
-  const records = resources.harvesterImpls?.records || [];
-  const implementations = records.length
-    ? records[0].implementations
+const extractHarvesterImpls = (resources, labelName) => {
+  const implementations = resources.length
+    ? resources[0].implementations
     : [];
   const results = implementations.map(i => ({
     value: i.type,
-    label: i.name,
+    label: labelName ? i[labelName] : i.name,
   }));
   results.unshift({ value: undefined, label: '' });
   return results;